+++ ptyrad/__init__.py
@@ -1,5 +1,5 @@
 # __init__.py is not strictly needed since Python 3.3 for the intepreter to recognize it as a package 
 # However, this is still included for clarity and for initialization purpose
 
-__version__ = "0.1.0-beta0.1" # 2024.0523
+__version__ = "0.1.0-beta1.1" # 2024.0531
 print(f"Importing ptyrad {__version__}")