+++ ptyrad/inputs/full_params_tBL_WSe2.py
@@ -1,6 +1,6 @@
 # tBL_WSe2
 
-ptycho_output_path = 'data/tBL_WSe2/Fig_1h_24.9mrad_Themis/1/roi1_Ndp128_step128/MLs_L1_p10_g128_pc0_noModel_updW100_mm_dpFlip_ud_T/Niter9000_v7.mat'
+ptycho_output_path = 'output/tBL-WSe2/20240529_full_N16384_dp128_flipT100_random32_p16_plr1e-4_oalr5e-4_oplr5e-4_slr1e-4_tlr0_1obj_12slice_dz1.0_kzf0.1_prev_probe_pos_sparse/model_iter0200.pt'#'data/tBL_WSe2/Fig_1h_24.9mrad_Themis/1/roi1_Ndp128_step128/MLs_L1_p10_g128_pc0_noModel_updW100_mm_dpFlip_ud_T/Niter9000_v7.mat'
 exp_CBED_path      = 'data/tBL_WSe2/Fig_1h_24.9mrad_Themis/1/data_roi1_Ndp128_step128_dp.hdf5'
 
 exp_params = {
@@ -11,7 +11,7 @@ exp_params = {
     'dx_spec'           : 0.1494,# Ang
     'defocus'           : 0, # Ang, positive defocus here refers to actual underfocus or weaker lens strength following Kirkland/abtem/ptychoshelves convention
     'c3'                : 0, # Ang, spherical aberration coefficients
-    'z_distance'        : 8, # Ang
+    'z_distance'        : 1, # Ang
     'Nlayer'            : 1,
     'N_scans'           : 16384,
     'N_scan_slow'       : 128,
@@ -19,10 +19,11 @@ exp_params = {
     'scan_step_size'    : 0.4290, # Ang
     'scan_flipT'        : None, # None for both 'simu' and loaded pos. Modify scan_flipT would change the image orientation. Expected input is [flipup, fliplr, transpose] just like PtychoShleves
     'scan_affine'       : None, # (scale, asymmetry, rotation, shear)
+    'scan_rand_std'     : 0.15, # None or scalar. Randomize the initial guess of scan position with Gaussian distributed displacement (std in px) to avoid raster grid pathology
     'obj_tilts'         : {'tilt_type':'all', 'init_tilts':[[0,0]]}, # (tilt_y,tilt_x) mrad, 'tilt_type' = 'all', 'each', or 'load_PtyRAD'
     'omode_max'         : 1,
     'omode_init_occu'   : {'occu_type':'uniform', 'init_occu':None},
-    'pmode_max'         : 10,
+    'pmode_max'         : 16,
     'pmode_init_pows'   : [0.02],
     'probe_permute'     : None,
     'cbeds_permute'     : None,
@@ -63,12 +64,13 @@ model_params = {
         'objp'            : 5e-4,
         'obj_tilts'       : 0, 
         'probe'           : 1e-4, 
-        'probe_pos_shifts': 1e-4}}
+        'probe_pos_shifts': 5e-4}}
 
 loss_params = {
-    'loss_single': {'state': True, 'weight': 1.0, 'dp_pow': 0.5},
-    'loss_pacbed': {'state': False, 'weight': 0.5, 'dp_pow': 0.2},
-    'loss_sparse': {'state': False, 'weight': 0.1, 'ln_order': 1},
+    'loss_single': {'state': False, 'weight': 1.0, 'dp_pow': 0.5},
+    'loss_poissn': {'state': True, 'weight': 1.0, 'dp_pow':1.0, 'eps':0},
+    'loss_pacbed': {'state': False, 'weight': 0.5, 'dp_pow': 0.2}, # Usually weight:0.5, dp_pow:0.2
+    'loss_sparse': {'state': True, 'weight': 0.01, 'ln_order': 1},
     'loss_simlar': {'state': False, 'weight': 1.0, 'obj_type':'both', 'scale_factor':[1,1,1], 'blur_std':1}
 }
 
@@ -78,7 +80,7 @@ constraint_params = {
     'fix_probe_int' : {'freq': 1},
     'obj_blur'      : {'freq': None, 'obj_type':'both', 'std':1},
     'kr_filter'     : {'freq': None,    'obj_type':'both', 'radius':0.15, 'width':0.05},
-    'kz_filter'     : {'freq': None,    'obj_type':'both', 'beta':1, 'alpha':1},
+    'kz_filter'     : {'freq': None,    'obj_type':'both', 'beta':0.1, 'alpha':1},
     'obja_thresh'   : {'freq': None, 'relax':0, 'thresh':[0.95**(1/25), 1.05**(1/25)]},
     'objp_postiv'   : {'freq': 1,    'relax':0},
     'tilt_smooth'   : {'freq': None, 'std':2}
@@ -91,14 +93,14 @@ def get_date(date_format = '%Y%m%d'):
     return date_str
 
 # Recon params
-NITER        = 20
-INDICES_MODE = 'sub'   # 'full', 'center', 'sub'
-BATCH_SIZE   = 128
+NITER        = 50
+INDICES_MODE = 'full'   # 'full', 'center', 'sub'
+BATCH_SIZE   = 32
 GROUP_MODE   = 'random' # 'random', 'sparse', 'compact' # Note that 'sparse' for 256x256 scan could take more than 10 mins on CPU. PtychoShelves automatically switch to 'random' for Nscans>1e3
-SAVE_ITERS   = 5        # scalar or None
+SAVE_ITERS   = 10        # scalar or None
 
 # Output folder and pre/postfix, note that the needed / and _ are automatically generated
 output_dir   = 'output/tBL-WSe2'
 prefix       = get_date(date_format='%Y%m%d')
-postfix      = 'altas'
+postfix      = ''
 fig_list     = ['loss', 'forward', 'probe_r_amp', 'probe_k_amp', 'probe_k_phase', 'pos'] # 'loss', 'forward', 'probe_r_amp', 'probe_k_amp', 'probe_k_phase', 'pos', 'tilt', or 'all' for all the figures